requirements.txt
pytest==2.9.0
pytest-bdd
selenium
jsonpickle
pony
pymysql
mysql-connector-python-rf
lxml-4.6.2-cp39-cp39-win_amd64.whl
allure-pytest